requirements.txt
colorama>=0.4.4
psutil>=5.9.0
wmi>=1.5.1
Pillow>=9.0.0
rich>=12.0.0 